chore: move settings to inner components
