+++ src/screens/settings/components/ChangeEmailSetting.tsx
@@ -1,10 +1,10 @@
 import { useNavigation } from '@react-navigation/native';
 import React from 'react';
 
-import Card from '../../components/Card';
-import ClickableText from '../../components/ClickableText';
-import { ServicesNavigationProp } from '../../navigation/types';
-import { fontSize } from '../../utils/texts';
+import Card from '../../../components/Card';
+import ClickableText from '../../../components/ClickableText';
+import { ServicesNavigationProp } from '../../../navigation/types';
+import { fontSize } from '../../../utils/texts';
 
 export default function ChangeEmailSetting() {
   const navigation = useNavigation<ServicesNavigationProp>();
