+++ src/screens/settings/components/ResetIntroSetting.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
-import ClickableText from '../../components/ClickableText';
-import { useAppDispatch } from '../../hooks';
-import { setIntroViewed } from '../../redux/reducers/settingsSlice';
-import { fontSize } from '../../utils/texts';
+import ClickableText from '../../../components/ClickableText';
+import { useAppDispatch } from '../../../hooks';
+import { setIntroViewed } from '../../../redux/reducers/settingsSlice';
+import { fontSize } from '../../../utils/texts';
 
 const ResetIntroSetting = () => {
   const dispatch = useAppDispatch();
