+++ src/screens/settings/components/ToggleSignNotification.tsx
@@ -2,13 +2,13 @@ import { AntDesign } from '@expo/vector-icons';
 import React from 'react';
 import { Linking, StyleSheet, Switch, ToastAndroid, TouchableOpacity, View } from 'react-native';
 
-import { cache } from '../../cache/smartCache';
-import Text from '../../components/Text';
-import { useAppDispatch, useAppSelector, useGlobalStyles } from '../../hooks';
-import { setSignNotification } from '../../redux/reducers/settingsSlice';
-import { registerSignsFetchTask, unregisterBackgroundFetchAsync } from '../../tasks/signs/signs';
-import { NOTIFICATION_GUIDE_URL } from '../../utils';
-import { fontSize } from '../../utils/texts';
+import { cache } from '../../../cache/smartCache';
+import Text from '../../../components/Text';
+import { useAppDispatch, useAppSelector, useGlobalStyles } from '../../../hooks';
+import { setSignNotification } from '../../../redux/reducers/settingsSlice';
+import { registerSignsFetchTask, unregisterBackgroundFetchAsync } from '../../../tasks/signs/signs';
+import { NOTIFICATION_GUIDE_URL } from '../../../utils';
+import { fontSize } from '../../../utils/texts';
 
 const styles = StyleSheet.create({
   cardView: {
@@ -21,7 +21,7 @@ const styles = StyleSheet.create({
 
 const ToggleSignNotification = () => {
   const dispatch = useAppDispatch();
-  const signNotification = useAppSelector((state) => state.settings.signNotification);
+  const signNotification = useAppSelector((state) => state.settings.config.signNotificationEnabled);
   const globalStyles = useGlobalStyles();
   const { isDemo, isOfflineMode } = useAppSelector((state) => state.auth);
